Refactor rest endpoints "navigation/children"
