+++ library/App.php
@@ -116,6 +116,8 @@ class App
          * Api
          */
         RestApiEndpointsRegistry::add(new \Municipio\Api\Media\Sideload());
+        RestApiEndpointsRegistry::add(new \Municipio\Api\Navigation\Children());
+        RestApiEndpointsRegistry::add(new \Municipio\Api\Navigation\ChildrenRender());
 
         /**
          * Customizer
